+++ include/hep/mc/mpi_helper.hpp
@@ -3,7 +3,7 @@
 
 /*
  * hep-mc - A Template Library for Monte Carlo Integration
- * Copyright (C) 2013  Christopher Schwan
+ * Copyright (C) 2013-2014  Christopher Schwan
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -80,42 +80,6 @@ inline void mpi_advance_generator_after(
 
 	generator.discard(after * usage);
 }
-
-}
-
-namespace hep
-{
-
-/// \addtogroup mpi_helper
-/// @{
-
-/**
- * This function returns true if the MPI functions should use a single random number generator. If
- * this is the case the generator of each process discards an approriate amount of random numbers to
- * ascertain independent numbers are used. This make sure the result is independent of the number of
- * processes. However, since each process consumes as many random numbers as calls are made in
- * total, this may slow down the integration if the integrand evaluates too fast. In this case a
- * faster random number generator can be used.
- */
-inline bool& mpi_single_generator()
-{
-	static bool use_single_generator = false;
-
-	return use_single_generator;
-}
-
-/**
- * If `enabled` is true, the MPI function's results are independent of the number of processes used
- * to integrate.
- *
- * \see mpi_single_generator
- */
-inline void mpi_single_generator(bool enabled)
-{
-	mpi_single_generator() = enabled;
-}
-
-/// @}
 
 }